add tap_dance files
