+++ users/scotto/scotto.h
@@ -16,6 +16,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include QMK_KEYBOARD_H
 
+#ifdef TAP_DANCE_ENABLE
+#   include "tap_dance.h"
+#endif
+
 #define _________________QWERTY1__________________      KC_Q,         KC_W, KC_E, KC_R, KC_T, KC_Y, KC_U, KC_I,    KC_O,   KC_P
 #define _________________QWERTY2__________________      KC_A,         KC_S, KC_D, KC_F, KC_G, KC_H, KC_J, KC_K,    KC_L,   KC_BSPC
 #define _________________QWERTY3a_________________      LSFT_T(KC_Z), KC_X, KC_C, KC_V, KC_B, KC_N, KC_M, KC_COMM, KC_DOT, RSFT_T(KC_SLSH)
